plat/native/arch/x86_64: Adapt to `idx`/`id`/`auxsp` updates
`idx`/`id` and related functionalities have had their ownership
renounced to `libukpcpuvar` so remove related implementations.

What is more, define the new PAL per-CPU variable for `auxsp`
and update `uk_plat_native_lcpu_init` to `uk_plat_native_except_init`.

Finally, an immediately visible great benefit to this is being able
remove any LCPU dependencies (except for get_idx_in_except which is
unfortunately needed until we get proper swapgs semantics and the
guarantee that the per-CPU register holds a kernel-owned value
during exception handling).
With this occasion, separate the lcpu_pm_ops registration and related
functionalities into its own source file, away from the other PAL
specific ones and define a new CONFIG_LIBUKPLAT_NATIVE_LCPU_PM,
selectable by platforms.
