+++ plat/kvm/Config.uk
@@ -18,6 +18,7 @@ menuconfig PLAT_KVM
 	select HAVE_PAGING
 	select LIBUKPLAT_NATIVE_PAL
 	select LIBUKPLAT_NATIVE_PAGING if LIBUKPAGING
+	select LIBUKPLAT_NATIVE_LCPU_PM if LIBUKLCPU
 	select HAVE_INTCTLR
 	select HAVE_APIC if ARCH_X86_64
 	select LIBUKINTCTLR_XPIC if ARCH_X86_64
