+++ plat/native/Makefile.uk
@@ -30,6 +30,9 @@ LIBUKPLAT_NATIVE_SRCS-$(CONFIG_LIBUKPLAT_NATIVE_PAGING) += $(LIBUKPLAT_NATIVE_BA
 LIBUKPLAT_NATIVE_SRCS-$(CONFIG_LIBUKPLAT_NATIVE_PAGING) += $(LIBUKPLAT_NATIVE_BASE)/arch/x86_64/pt.c|isr
 LIBUKPLAT_NATIVE_SRCS-$(CONFIG_LIBUKPLAT_NATIVE_ECTX) += $(LIBUKPLAT_NATIVE_BASE)/arch/x86_64/ectx.c|isr
 LIBUKPLAT_NATIVE_SRCS-$(CONFIG_LIBUKPLAT_NATIVE_EXCEPT) += $(LIBUKPLAT_NATIVE_BASE)/arch/x86_64/except.c|isr
-LIBUKPLAT_NATIVE_SRCS-y += $(LIBUKPLAT_NATIVE_BASE)/arch/x86_64/lcpu.c
+LIBUKPLAT_NATIVE_SRCS-$(CONFIG_LIBUKPLAT_NATIVE_SYSCTX) += $(LIBUKPLAT_NATIVE_BASE)/arch/x86_64/sysctx_auxsp.c
+ifeq ($(CONFIG_LIBUKLCPU), y)
+LIBUKPLAT_NATIVE_SRCS-$(CONFIG_LIBUKPLAT_NATIVE_LCPU_PM) += $(LIBUKPLAT_NATIVE_BASE)/arch/x86_64/lcpu_pm.c
+endif
 endif
 endif
