+++ plat/native/arch/x86_64/include/uk/plat/native/arch/auxsp.h
@@ -9,6 +9,12 @@
 
 #include <uk/arch/types.h>
 #include <uk/config.h>
+#include <uk/pcpuvar.h>
+
+#if CONFIG_LIBUKPLAT_NATIVE_AUXSP
+#define UK_PLAT_NATIVE_AUXSP_SYM					\
+	uk_plat_native_auxsp
+#endif /* CONFIG_LIBUKPLAT_NATIVE_AUXSP */
 
 #if !__ASSEMBLY__
 #ifdef __cplusplus
@@ -16,29 +22,9 @@ extern "C" {
 #endif
 
 #if CONFIG_LIBUKPLAT_NATIVE_AUXSP
-/**
- * Get current auxiliary stack pointer.
- * The auxiliary stack is used for exception handling.
- *
- * @return Auxiliary stack pointer value
- */
-__uptr uk_plat_native_get_auxsp(void);
+extern __uk_pcpuvar __uptr uk_plat_native_auxsp;
 
-/**
- * Set auxiliary stack pointer.
- * Configures the stack used for exception handling.
- *
- * @param auxsp  New auxiliary stack pointer
- */
 void uk_plat_native_set_auxsp(__uptr auxsp);
-
-/**
- * Get auxiliary stack pointer during exception handling.
- * ISR-safe variant for use within exception handlers.
- *
- * @return Current auxiliary stack pointer
- */
-__isr __uptr uk_plat_native_get_auxsp_in_except(void);
 #endif /* CONFIG_LIBUKPLAT_NATIVE_AUXSP */
 
 #ifdef __cplusplus
